refactor sanityClient
