+++ __tests__/applications.test.ts
@@ -1,4 +1,4 @@
-import { createApplication } from "@/app/actions/actions";
+import { createApplication } from "@/app/actions/sanityClient";
 import { validatePlanningParams } from "@/app/actions/validator";
 import handler from "@/app/api/applications/route";
 import { verifyApiKey } from "../src/app/lib/apiKey";
