--- src/app/(main)/concern-info.tsx
@@ -1,8 +1,0 @@
-import { getGlobalContent } from "../actions/actions";
-
-async function ConcernInfo() {
-  const globalConfig: any = getGlobalContent();
-  return <p className="govuk-body">{globalConfig?.concernContent}</p>;
-}
-
-export default ConcernInfo;
