+++ src/app/(main)/concern-info/page.tsx
@@ -1,0 +1,9 @@
+// import { getGlobalContent } from "../actions/actions";
+import { getGlobalContent } from "../../actions/sanityClient";
+
+async function ConcernInfo() {
+  const globalConfig: any = await getGlobalContent();
+  return <p className="govuk-body">{globalConfig?.concernContent}</p>;
+}
+
+export default ConcernInfo;
