+++ src/app/actions/client.ts
@@ -1,16 +1,16 @@
-import { createClient } from "@sanity/client";
-import imageUrlBuilder from "@sanity/image-url";
+// import { createClient } from "@sanity/client";
+// import imageUrlBuilder from "@sanity/image-url";
 
-export const client = createClient({
-  projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
-  dataset: process.env.NEXT_PUBLIC_SANITY_DATASET,
-  useCdn: false,
-  apiVersion: "2023-11-15",
-  token: process.env.NEXT_PUBLIC_SANITY_SECRET_TOKEN,
-});
+// export const client = createClient({
+//   projectId: process.env.NEXT_PUBLIC_SANITY_PROJECT_ID,
+//   dataset: process.env.NEXT_PUBLIC_SANITY_DATASET,
+//   useCdn: false,
+//   apiVersion: "2023-11-15",
+//   token: process.env.NEXT_PUBLIC_SANITY_SECRET_TOKEN,
+// });
 
-const builder = imageUrlBuilder(client);
+// const builder = imageUrlBuilder(client);
 
-export function urlFor(source: any) {
-  return builder.image(source);
-}
+// export function urlFor(source: any) {
+//   return builder.image(source);
+// }
