+++ src/components/feedback-information/index.tsx
@@ -5,7 +5,8 @@ import { Button } from "@/components/button";
 import { ArrowIcon } from "../../../public/assets/icons";
 import { useEffect, useState } from "react";
 import { getLocalStorage } from "@/app/lib/application";
-import { getGlobalContent } from "@/app/actions/actions";
+// import { getGlobalContent } from "@/app/actions/actions";
+import { getGlobalContent } from "../../app/actions/sanityClient";
 
 function FeedbackInformation({
   onChangeQuestion,
