+++ src/components/process/index.tsx
@@ -7,7 +7,9 @@ import {
   applicationStageMessage,
 } from "@/app/lib/application";
 import { useEffect, useState } from "react";
-import { getGlobalContent } from "@/app/actions/actions";
+// import { getGlobalContent } from "@/app/actions/actions";
+import { getGlobalContent } from "../../app/actions/sanityClient";
+
 function Process({
   data,
   consultationDeadline,
